% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib2df.R
\name{bib2df}
\alias{bib2df}
\title{Parse a BibTeX file to a tibble.}
\usage{
bib2df(file)
}
\arguments{
\item{file}{character, path to a .bib file.}
}
\value{
A tibble.
}
\description{
The BibTeX file is read, parsed, tidied and written to a tibble.
}
\details{
Uses the internal \code{bib2df_read()}, \code{bib2df_gather()} and \code{bib2df_tidy()}. No magic, just a wrapper.
}
\examples{
path <- system.file("extdata", "biblio.bib", package = "bib2df")
bib2df(path)
}
\seealso{
\code{\link{df2bib}}
}
\author{
Philipp Ottolinger
}
