% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib2f_fields_list.R
\docType{data}
\name{bib2df_fields_list}
\alias{bib2df_fields_list}
\title{Supported Fields for Parsing}
\description{
This document lists all the fields that can be parsed by default in `bib2df`.
}
\section{Fields}{

\itemize{
  \item CATEGORY
  \item BIBTEXKEY
  \item ADDRESS
  \item ANNOTE
  \item AUTHOR
  \item BOOKTITLE
  \item CHAPTER
  \item CROSSREF
  \item EDITION
  \item EDITOR
  \item HOWPUBLISHED
  \item INSTITUTION
  \item JOURNAL
  \item KEY
  \item MONTH
  \item NOTE
  \item NUMBER
  \item ORGANIZATION
  \item PAGES
  \item PUBLISHER
  \item SCHOOL
  \item SERIES
  \item TITLE
  \item TYPE
  \item VOLUME
  \item YEAR
  \item AFFILIATION
  \item ABSTRACT
  \item DOI
  \item EID
  \item FILE
  \item CONTENTS
  \item COPYRIGHT
  \item COLLABORATOR
  \item ISBN
  \item ISSN
  \item KEYWORDS
  \item LANGUAGE
  \item LOCATION
  \item LCCN
  \item MRNUMBER
  \item PMC
  \item PMID
  \item PRICE
  \item SIZE
  \item TRANSLATOR
  \item URL
  \item AUTHOR_KEYWORDS
}
}

