% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib2df_read.R
\name{bib2df_read}
\alias{bib2df_read}
\title{Read a .bib file.}
\usage{
bib2df_read(file)
}
\arguments{
\item{file}{character, path to a .bib file.}
}
\value{
A character vector.
}
\description{
\code{bib2df_read()} reads a .bib file to a character vector.
}
\details{
\code{bib2df_read()} uses \code{base::readLines()} to read from a local .bib file. UTF-8 compatibility is ensured.
}
\examples{
path <- system.file("extdata", "biblio.bib", package = "bib2df")
bib <- bib2df_read(path)
bib
}
\author{
Philipp Ottolinger
}

