% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib2df_tidy.R
\name{bib2df_tidy}
\alias{bib2df_tidy}
\title{Tidy up data frame}
\usage{
bib2df_tidy(bib, separate_names = FALSE)
}
\arguments{
\item{bib}{(data.frame) dataframe of .bib document}

\item{separate_names}{(logical) whether to separate names in author and editor fields}
}
\description{
Trims ws of all fields, removes extra braces in double bracing situations,
and tidy up editor, author, and year fields.
}
