% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2bib.R
\name{df2bib}
\alias{df2bib}
\title{Export a BibTeX tibble to a .bib file.}
\usage{
df2bib(x, file)
}
\arguments{
\item{file}{character, path to a .bib file.}

\item{a}{tibble returned by \code{\link{df2bib}}.}
}
\value{
\code{file} as a character string, invisibly.
}
\description{
The BibTeX file is read, parsed, tidied and written to a tibble.
}
\examples{
path <- system.file("extdata", "biblio.bib", package = "bib2df")
bib <- bib2df(path)
df2bib(bib, bib2 <- tempfile())
identical(bib, bib2df(bib2))
}
\seealso{
\code{\link{bib2df}}
}
\author{
Thomas J. Leeper
}
