% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib2df_gather.R
\name{parse_entry}
\alias{parse_entry}
\title{Parse a single BibTeX entry}
\usage{
parse_entry(entry, extra_fields)
}
\arguments{
\item{entry}{(char) a BibTeX entry}

\item{extra_fields}{(char) A list of additional fields to parse, passed from `bib2df()`}
}
\description{
Parse a single BibTeX entry
}
