% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib2df-package.R
\docType{package}
\name{bib2df-package}
\alias{bib2df-package}
\title{bib2df: Parse a BibTeX file to \code{data.frame}}
\description{
This package provides functions to parse and write BibTeX files. BibTeX files
can be parsed to \code{data.frame} to make them accessible with popular tools like
\code{dplyr}, \code{tidyr}, \code{ggplot2} and many more.
}
\details{
BibTeX entries represented in a \code{data.frame} can be altered in a familiar way and
written back to a valid BibTeX file.

To learn more about \code{bib2df}, start with the vignettes:
 \code{browseVignettes(package = "bib2df")}
}
