% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib2df.R
\name{bib2df}
\alias{bib2df}
\title{Parse a BibTeX file to a \code{data.frame}}
\usage{
bib2df(file, separate_names = FALSE)
}
\arguments{
\item{file}{character, path to a .bib file.}

\item{separate_names}{logical, should authors' and editors' names be separated into first and given name?}
}
\value{
A \code{data.frame}.
}
\description{
The BibTeX file is read, parsed, tidied and written to a \code{data.frame}
}
\details{
For simplicity \code{bib2df()} unifies the reading, parsing and tidying of a BibTeX file while being aware of a standardized output format, different BibTeX styles and missing values in the BibTeX file.
}
\examples{
path <- system.file("extdata", "biblio.bib", package = "bib2df")
bib2df(path)
}
\seealso{
\code{\link{df2bib}}
}
\author{
Philipp Ottolinger
}
