% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2bib.R
\name{df2bib}
\alias{df2bib}
\title{Export a BibTeX \code{data.frame} to a .bib file}
\usage{
df2bib(x, file)
}
\arguments{
\item{x}{\code{data.frame}, returned by \code{\link{df2bib}}.}

\item{file}{character, path to a .bib file.}
}
\value{
\code{file} as a character string, invisibly.
}
\description{
The BibTeX \code{data.frame} is written to a .bib file
}
\examples{
path <- system.file("extdata", "biblio.bib", package = "bib2df")
bib <- bib2df(path)
df2bib(bib, bib2 <- tempfile())
identical(bib, bib2df(bib2))
}
\seealso{
\code{\link{bib2df}}
}
\author{
Thomas J. Leeper
}
